Moved tests from importer.py to test_importer.py (#215)
Closes #25.
