+++ tests/test_importer.py
@@ -36,3 +36,65 @@ class TestImporter(unittest.TestCase):
         self.assertEqual(ing.amount, 2)
         self.assertEqual(ing.unit, "cups")
         self.assertEqual(ing.item, "water")
+
+
+class ImporterTest(unittest.TestCase):
+
+    def setUp(self):
+        self.importer = importer.Importer()
+
+    def test_parse_simple_yields(self):
+        assert self.importer.parse_yields("3 cups") == (3, "cups")
+        assert self.importer.parse_yields("7 servings") == (7, "servings")
+        assert self.importer.parse_yields("12 muffins") == (12, "muffins")
+        assert self.importer.parse_yields("10 loaves") == (10, "loaves")
+
+    def test_parse_complex_yields(self):
+        assert self.importer.parse_yields("Makes 12 muffins") == (12, "muffins")
+        assert self.importer.parse_yields("Makes 4 servings") == (4, "servings")
+        assert self.importer.parse_yields("Serves 7") == (7, "servings")
+
+    def test_parse_fractional_yields(self):
+        assert self.importer.parse_yields("Makes 4 3/4 muffins") == (4.75, "muffins")
+        assert self.importer.parse_yields("Makes 4 3/4") == (4.75, "servings")
+        assert self.importer.parse_yields("19/4") == (4.75, "servings")
+        assert self.importer.parse_yields("Makes 19/4") == (4.75, "servings")
+
+    @unittest.expectedFailure
+    def test_failed_parsing_fractional_yields(self):
+        assert self.importer.parse_yields("Makes 19/4") == (4.75, "muffins")
+
+
+class RatingConverterTest(unittest.TestCase):
+
+    def setUp(self):
+
+        class FakeDB:
+
+            recs = dict([(n, {}) for n in range(20)])
+
+            def get_rec(self, n):
+                return n
+
+            def modify_rec(self, n, d):
+                for attr, val in list(d.items()):
+                    self.recs[n][attr] = val
+
+        self.db = FakeDB()
+
+    def test_automatic_converter(self):
+        rc = importer.RatingConverter()
+        tests = [("good", 6), ("Great", 8), ("Excellent", 10), ("poor", 2), ("okay", 4)]
+        for n, (rating, number) in enumerate(tests):
+            rc.add(n, rating)
+            self.db.recs[n]["rating"] = rating
+        rc.do_conversions(self.db)
+        print("Conversions: ")
+        for n, (rating, number) in enumerate(tests):
+            print("Converted", rating, "->", self.db.recs[n]["rating"])
+            self.assertEqual(number, self.db.recs[n]["rating"])
+
+    def test_string_to_rating_converter(self):
+        assert importer.string_to_rating("4/5 stars") == 8
+        assert importer.string_to_rating("3 1/2 / 5 stars") == 7
+        assert importer.string_to_rating("4/10 stars") == 4
